Support log message filtering
Sometimes, especially when processing a large number of events /
requests, applications can go wild and start logging the same stuff
over and over again. It is desirable to filter messages in these
cases, so as to avoid filling up hard drives or to avoid exceeding
log aggregation limits. In other cases, sensitive information
should be filtered out.

This commit adds basic facilities for log message filtering. These
allow the registration of log message filters which are applied
before (potentially) expensive formatting and downstreaming
tasks are executed. This is similar to other logging libraries
such as logback in the Java world.