Remove check for `undefined` in `isPinoLine`
According to https://tools.ietf.org/html/rfc7159#section-3
the only values other than arrays and objects that are allowed in
JSON are `true`, `false`, and `null`. So we don't need to check
for `undefined`. Further, we don't need an explicit check for
`true` or `false` because they are actual objects that have
the `hasOwnProperty` function.