Fix addLevel bugs
Changes made:

- Move lscache to Pino.prototype
- When addLevel is called:
  - First time: Create local copies of lscache
    and levels (as these are about to be modified), on the instance
  - Add the new level as a function on the instance
- When creating a new child, if new levels have been added to the parent,
  copy these over to the child.

NOTE: In order to not make copies for every child, if the parent has
custom levels, the same levels and lscache objects are used as for the
parent. This means that levels added to the parent will be visible on
the levels property for children as well (although no new log methods
will be created on the children).
It also means that levels added to the child, will be visible on
the levels property for the parent as well (although no new log methods
will be created on the parent).

I don't see this as a problem as addLevel will most likely be called on
the first instance created, before any children are created.