fix throw when `redact.remove` acts on chindings
The previous commit holds the failing test case.
This commit fixes the issue, passing the test.

The issue was that using `redact.remove` on a key that was
subsequently used in a child binding, was cause logging on
the child logger to produce JSON with `"key":undefined` in it.

While this doesn't crash the logger, the JSON is invalid and therefore
will break many recieving systems, at best causing logs to be missed.

In the test suite, the output is `JSON.parse`d which throws due to the
invalid output.