fix(redaction): fix corner cases with paths (#732)
In particular, this supports single-item paths where the first item is
bracketed (e.g. `{ redact: ["['some-key']"] }`) and paths where one of
the segments is the empty string [in brackets] (e.g., `{ redact:
["a[''].x"] }`).

Conceptually, the change is quite simple: it uses the namespace rather,
than the string, even for single-segment paths, since the namespace has
the brackets + quotes removed. It also changes the namespace `.replace`
regex to allow 0 character (i.e., `.*` instead of `.+`) between the
quotes.